Fix cursor jumping to beginning of buffer
If `M-x srfi` was done with the "*SRFI*" buffer being displayed in a
window other than the currently active one, the cursor would jump all
the way to the top left corner of the "*SRFI*" buffer. Fix the bug so
it jumps to the line listing the first SRFI instead.