Revert "Shorten using `dotimes'."
While `dotimes` is indeed a better way to write the loop, it causes
the latest version of Emacs (26.3) to emit a spurious byte-compiler
warning about the lexical variable `n` being unused even though we are
using it. Naming the variable `_n` (the `_` prefix meaning "unused")
causes a different compiler warning.

Compiling without warnings one important quality marker for released
software, so let's keep it even at the expense of clumsier code until
the byte-compiler is fixed.

This reverts commit fefbfcaeaff953853d7492482c46bf0255f2ac12.