Revert adding srfi-browser-function variable
The standard `browse-url-browser-function` variable can be set to a
list of (URL-REGEXP . BROWSE-FUNCTION) pairs for precise control over
which browser to use for what URLs. Hence we don't need tailor-made
browser control in the SRFI package.

One benefit of the fine-grained regexp matching is that the SRFI home
page and discussions can be set to open in a graphical browser while
the SRFI documents themselves can open inside Emacs in a text browser.